Code cleanup in Reqnroll.Parser
