+++ Reqnroll.Generator/Generation/UnitTestFeatureGenerator.cs
@@ -14,7 +14,7 @@ using Reqnroll.Generator.CodeDom;
 using Reqnroll.Generator.UnitTestConverter;
 using Reqnroll.Generator.UnitTestProvider;
 using Reqnroll.Parser;
-using Reqnroll.Parser.CucmberMessageSupport;
+using Reqnroll.Parser.CucumberMessages;
 using Reqnroll.Tracing;
